simplify build
rollup now builds in one step with multiple entry points to the app.
This allows for efficient code splitting in chunks. For this to work
properly with the cjs format, the entry points need to be required.
Therefore the html pages don't include via script tag but through
require calls.